requirements.txt
langchain
streamlit
streamlit-chat
python-dotenv
streamlit-audiorec
asyncio
yandex-speechkit
aiohttp